backend/requirements.txt
fastapi
uvicorn
pydantic
ragas
sqlalchemy
pandas
python-dotenv
openai
numpy
langchain-openai
openpyxl
langchain-community
datasets
nest-asyncio
langchain-huggingface
chromadb
pytest
pytest-asyncio
